refactor: scatter chart 리팩토링(#175)
