+++ src/features/game/components/ranking/common/chart/CustomIndicator.tsx
